+++ src/dexterity/localroles/interfaces.py
@@ -4,4 +4,19 @@ from zope.interface import Interface
 
 
 class IDexterityLocalRoles(Interface):
-    """Specific layer for the package"""
+    """ Specific layer for the package """
+
+
+class ILocalRolesRelatedSearchUtility(Interface):
+    """ Interface for related local roles business """
+
+    def get_objects(self, context):
+        """ Get related objects.
+
+        :param context: original type instance for which a related configuration is defined
+        :type context: type instance object
+
+        :returns: a list of objects
+        :rtype: list
+        """
+        pass